requirements/base.txt

Django==1.10.5
django-allauth==0.30.0
django-braces==1.11.0
django-environ==0.4.1
django-extensions==1.7.6
django-model-utils==2.6.1
Pillow==4.0.0
psycopg2==2.6.2
pytz==2016.10
raven==5.32.0
django-solo==1.1.2
django-rest-swagger==2.1.1
djangorestframework==3.5.3
django-rest-auth==0.9.1
django-filter==1.0.1
django-cors-headers==2.1.0
geopy==1.11.0
